Remove floats outside of mathutils
Floats are one of the lesser used types. Addtionally, by default, all of the WPILib, CTRE, and NavX libraries use doubles. By using doubles, we don't have to cast numbers to floats all the time, which is much cleaner in my opinion.

Also, if we get memory issues from this, we have bigger issues.